requirements.txt
dlib==19.21.0
matplotlib==3.3.1
imageio==2.9.0
Flask==1.1.2
Werkzeug==1.0.1
scipy==1.5.2
numpy==1.19.1
Pillow==7.2.0
tensorflow==2.3.0
gunicorn==20.0.4
